删除 view
