+++ lite_mms/portal/delivery/__init__.py
@@ -12,8 +12,6 @@ from lite_mms.portal.delivery import views, ajax
 
 from lite_mms.basemain import data_browser, nav_bar
 
-from view import delivery_session_view, delivery_task_view, consigment_model_view, consigment_product_model_view
-
 extra_params = {
     "list_view": {
         "nav_bar": nav_bar,
@@ -24,7 +22,7 @@ extra_params = {
         "titlename": u"编辑发货单",
     }
 }
-data_browser.register_model_view(consigment_model_view, consignment_page, extra_params)
+data_browser.register_model_view(views.consigment_model_view, consignment_page, extra_params)
 
 extra_params = {
     "form_view": {
@@ -32,22 +30,22 @@ extra_params = {
         "titlename": u"编辑发货单项",
     }
 }
-data_browser.register_model_view(consigment_product_model_view, consignment_page, extra_params)
+data_browser.register_model_view(views.consigment_product_model_view, consignment_page, extra_params)
 
 from lite_mms.apis import delivery
 get_customer_list = delivery.get_store_bill_customer_list
 
-data_browser.register_model_view(delivery_session_view, delivery_page,
+data_browser.register_model_view(views.delivery_session_view, delivery_page,
                                  extra_params={"list_view": {"nav_bar": nav_bar, "titlename": u"发货会话列表"},
                                                "create_view": {"nav_bar": nav_bar, "titlename": u"新建发货会话", "get_customer_list": get_customer_list},
                                                "form_view": {"nav_bar": nav_bar, "titlename": u"编辑发货会话"}})
 
-data_browser.register_model_view(delivery_task_view, delivery_page,
+data_browser.register_model_view(views.delivery_task_view, delivery_page,
                                  extra_params={"form_view": {"nav_bar": nav_bar, "titlename": u"编辑任务会话"}})
 
 @consignment_page.route("/")
 def index():
-    return redirect(consigment_model_view.url_for_list(order_by="create_time", desc=1))
+    return redirect(views.consigment_model_view.url_for_list())
 
 @consignment_page.before_request
 @login_required
